Widen Xi emitter registers

+++ src/jit/xm_codegen_internal.h
@@ -195,7 +195,7 @@ typedef struct {
     uint32_t suspend_smap_ids[XM_MAX_SUSPEND_ENTRIES];      // stack map id at each suspend point
     uint8_t suspend_result_regs[XM_MAX_SUSPEND_ENTRIES];  // physical register for await result per
                                                           // suspend_id
-    int16_t suspend_result_bc_slots[XM_MAX_SUSPEND_ENTRIES];  // bc_slot of await result vreg per
+    int32_t suspend_result_bc_slots[XM_MAX_SUSPEND_ENTRIES];  // bc_slot of await result vreg per
                                                               // suspend_id
     int32_t suspend_result_tag_offs[XM_MAX_SUSPEND_ENTRIES];  // vreg_runtime_tags offset per
                                                               // suspend_id (-1=none)
